add a new mixin ImageGLRenderable
